+++ src/common/map_loader.c
@@ -975,7 +975,7 @@ BOX TRACING
 */
 
 // 1/32 epsilon to keep floating point happy
-#define	DIST_EPSILON	(0.03125)
+#define	DIST_EPSILON	0.03125f
 
 vec3_t	trace_start, trace_end;
 vec3_t	trace_mins, trace_maxs;
@@ -1269,9 +1269,9 @@ void Map_RecursiveHullCheck (int32_t num, float p1f, float p2f, vec3_t p1, vec3_
 		if (trace_ispoint)
 			offset = 0;
 		else
-			offset = fabs(trace_extents[0]*plane->normal[0]) +
-				fabs(trace_extents[1]*plane->normal[1]) +
-				fabs(trace_extents[2]*plane->normal[2]);
+			offset = fabsf(trace_extents[0]*plane->normal[0]) +
+				fabsf(trace_extents[1]*plane->normal[1]) +
+				fabsf(trace_extents[2]*plane->normal[2]);
 	}
 
 	// see which sides we need to consider
@@ -1289,14 +1289,14 @@ void Map_RecursiveHullCheck (int32_t num, float p1f, float p2f, vec3_t p1, vec3_
 	// put the crosspoint DIST_EPSILON pixels on the near side
 	if (t1 < t2)
 	{
-		idist = 1.0/(t1-t2);
+		idist = 1.0f/(t1-t2);
 		side = 1;
 		frac2 = (t1 + offset + DIST_EPSILON)*idist;
 		frac = (t1 - offset + DIST_EPSILON)*idist;
 	}
 	else if (t1 > t2)
 	{
-		idist = 1.0/(t1-t2);
+		idist = 1.0f/(t1-t2);
 		side = 0;
 		frac2 = (t1 - offset - DIST_EPSILON)*idist;
 		frac = (t1 + offset + DIST_EPSILON)*idist;
